+++ src/game/data/spriteType.ts
@@ -1,5 +1,7 @@
 
 export const SpriteType = {
+    LIFE: 4,
+    MAGIC: 5,
     KEY: 6,
     KASHES: 18,
     ZLITOS: 19,