+++ src/game/state.ts
@@ -1,24 +1,40 @@
 import { omit } from 'lodash';
 import { getLanguageConfig } from '../lang';
 
-export function createState() {
+export interface GameConfig {
+    displayText: boolean;
+    musicVolume: number;
+    soundFxVolume: number;
+    voiceVolume: number;
+}
+
+export interface GameState {
+    config: GameConfig;
+    hero: any;
+    chapter: number;
+    flags: any;
+    save: Function;
+    load: Function;
+}
+
+export function createState(): GameState {
     return {
         config: Object.assign({
             displayText: true,
-            musicVolume: 0.8,
+            musicVolume: 0.5,
             soundFxVolume: 0.5,
             voiceVolume: 1.0
         }, getLanguageConfig()),
         hero: {
             behaviour: 0,
             prevBehaviour: 0,
-            life: 50,
+            life: 200,
             money: 0,
             magic: 0,
             keys: 0,
             fuel: 0,
             pinguin: 0,
-            clover: { boxes: 2, leafs: 0 },
+            clover: { boxes: 2, leafs: 1 },
             magicball: { level: 0, strength: 0, bounce: 0 }
         },
         chapter: 0,