+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Version history
 
+### [v0.3.0-beta]
+_June 9th, 2020_
+
+#### Gameplay
+
+- Added extra bonus projectile motion with gravity
+- Added Zone life and magic points bonus
+- Added Behaviour Menu
+- Implemented escalators / conveyors
+
+#### VR
+
+- Added support for Firefox.
+- Fixed music playing twice in VR
+
+#### Replacement 3D models
+
+- Various 3D models added to replace the original brick layouts.
+- Allow to replace individual bricks
+
+#### Bug fixes
+
+- Various fixes: object sprite placements, zone scene transitions and 3D object bounding boxes
+
+-----------------------------
+
+### [v0.2.3](https://github.com/agrande/lba2remake/compare/v0.2.2...v0.2.3) Bugfixes
+_June 12th, 2020_
+
+Remove ability to edit metadata on main server
+
+-----------------------------
+
+### [v0.2.2](https://github.com/agrande/lba2remake/compare/v0.2.1...v0.2.2) Bugfixes
+_May 31th, 2020_
+
+Improve VR support and fix main theme audio quality
+
+-----------------------------
+
 ### [v0.2.1](https://github.com/agrande/lba2remake/compare/v0.2.0...v0.2.1) Bugfixes
 _May 30th, 2020_
 
@@ -119,8 +159,6 @@ _May 3rd, 2018_
 _May 1st, 2018_
 
 The inspector area lets you take a look at the remake's game engine internals, this is a useful feature for debugging or learning how the engine works.
-
-![Inspector](doc/images/inspector.png)
 
 A video tutorial is available [here](https://youtu.be/qgCQObDvZbI) (for developpers)